requirements.txt
APScheduler==3.6.3
astroid==2.11.3
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
df2gspread==1.0.4
dill==0.3.4
google-api-core==2.7.2
google-api-python-client==1.6.7
google-auth==1.33.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.53.0
gspread==3.7.0
httplib2==0.19.1
idna==3.2
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
numpy==1.21.1
oauth2client==4.1.3
oauthlib==3.1.1
packaging==21.0
pandas==1.3.1
platformdirs==2.5.2
pluggy==0.13.1
protobuf==3.17.3
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.13.7
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
tabulate==0.8.9
toml==0.10.2
tomli==2.0.1
tornado==6.1
typing-extensions==4.2.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.6
wrapt==1.14.0
